Interpolate `do` syntax closures into single _
This allows a single _ to be treated as the slot into which the closure
created by do syntax is inserted.

It's somewhat overloaded in that `@_ map(_, xs)` normally means the _
is the identity, while `@_ map(_, xs) do ...` would mean the same as
normal do syntax without the `_`, but conceptually we're marking a "slot
of map" to receive a closure so it could be acceptably consistent.

Also, it seems likely to be extremely convenient.